% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg.R
\name{rmse}
\alias{rmse}
\title{Calculate Root Mean Square Error (RMSE)}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector of actual values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
The RMSE value as a numeric scalar.
}
\description{
This function computes the RMSE between actual and predicted values.
}
\examples{
actual <- c(3, -0.5, 2, 7)
predicted <- c(2.5, 0.0, 2, 8)
rmse(actual, predicted)
}
